ev: remove EvMobsimListener
In order to simplify EV stats, move the CSV writing functionality to corresponding collector classes: ChargerPowerCollector and EnergyConsumptionCollector.
