+++ contribs/ev/src/main/java/org/matsim/contrib/ev/stats/ChargerPowerCollector.java
@@ -19,12 +19,18 @@
 
 package org.matsim.contrib.ev.stats;
 
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import org.apache.commons.csv.CSVFormat;
+import org.apache.commons.csv.CSVPrinter;
 import org.matsim.api.core.v01.Id;
+import org.matsim.contrib.ev.EvUnits;
 import org.matsim.contrib.ev.charging.ChargingEndEvent;
 import org.matsim.contrib.ev.charging.ChargingEndEventHandler;
 import org.matsim.contrib.ev.charging.ChargingStartEvent;
@@ -33,7 +39,12 @@ import org.matsim.contrib.ev.fleet.ElectricFleet;
 import org.matsim.contrib.ev.fleet.ElectricVehicle;
 import org.matsim.contrib.ev.infrastructure.Charger;
 import org.matsim.contrib.ev.infrastructure.ChargingInfrastructure;
+import org.matsim.core.controler.IterationCounter;
+import org.matsim.core.controler.OutputDirectoryHierarchy;
 import org.matsim.core.events.MobsimScopeEventHandler;
+import org.matsim.core.mobsim.framework.events.MobsimBeforeCleanupEvent;
+import org.matsim.core.mobsim.framework.listeners.MobsimBeforeCleanupListener;
+import org.matsim.core.utils.misc.Time;
 import org.matsim.vehicles.Vehicle;
 
 import com.google.common.base.Preconditions;
@@ -42,10 +53,17 @@ import com.google.inject.Inject;
 /*
  * created by jbischoff, 26.10.2018
  */
-public class ChargerPowerCollector implements ChargingStartEventHandler, ChargingEndEventHandler, MobsimScopeEventHandler {
+public class ChargerPowerCollector
+		implements ChargingStartEventHandler, ChargingEndEventHandler, MobsimScopeEventHandler, MobsimBeforeCleanupListener {
 
-	private final ChargingInfrastructure chargingInfrastructure;
-	private final ElectricFleet fleet;
+	@Inject
+	private OutputDirectoryHierarchy controlerIO;
+	@Inject
+	private IterationCounter iterationCounter;
+	@Inject
+	private ChargingInfrastructure chargingInfrastructure;
+	@Inject
+	private ElectricFleet fleet;
 
 	private record TimeCharge(double time, double charge) {
 	}
@@ -56,12 +74,6 @@ public class ChargerPowerCollector implements ChargingStartEventHandler, Chargin
 	}
 
 	private final List<ChargingLogEntry> logList = new ArrayList<>();
-
-	@Inject
-	public ChargerPowerCollector(ElectricFleet fleet, ChargingInfrastructure chargingInfrastructure) {
-		this.fleet = fleet;
-		this.chargingInfrastructure = chargingInfrastructure;
-	}
 
 	@Override
 	public void handleEvent(ChargingEndEvent event) {
@@ -81,7 +93,20 @@ public class ChargerPowerCollector implements ChargingStartEventHandler, Chargin
 		chargeBeginCharge.put(event.getVehicleId(), new TimeCharge(event.getTime(), ev.getBattery().getCharge()));
 	}
 
-	public List<ChargingLogEntry> getLogList() {
-		return logList;
+	@Override
+	public void notifyMobsimBeforeCleanup(MobsimBeforeCleanupEvent event) {
+		try (CSVPrinter csvPrinter = new CSVPrinter(
+				Files.newBufferedWriter(Paths.get(controlerIO.getIterationFilename(iterationCounter.getIterationNumber(), "chargingStats.csv"))),
+				CSVFormat.DEFAULT.withDelimiter(';')
+						.withHeader("ChargerId", "chargeStartTime", "chargeEndTime", "ChargingDuration", "xCoord", "yCoord",
+								"energyTransmitted_kWh"))) {
+			for (ChargerPowerCollector.ChargingLogEntry e : logList) {
+				double energyKWh = Math.round(EvUnits.J_to_kWh(e.transmitted_Energy()) * 10.) / 10.;
+				csvPrinter.printRecord(e.charger().getId(), Time.writeTime(e.chargeStart()), Time.writeTime(e.chargeEnd()),
+						Time.writeTime(e.chargeEnd() - e.chargeStart()), e.charger().getCoord().getX(), e.charger().getCoord().getY(), energyKWh);
+			}
+		} catch (IOException e) {
+			throw new RuntimeException(e);
+		}
 	}
 }
