+++ contribs/ev/src/main/java/org/matsim/contrib/ev/stats/EnergyConsumptionCollector.java
@@ -20,19 +20,39 @@
 
 package org.matsim.contrib.ev.stats;
 
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Paths;
 import java.util.HashMap;
 import java.util.Map;
 
+import org.apache.commons.csv.CSVFormat;
+import org.apache.commons.csv.CSVPrinter;
 import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.network.Link;
+import org.matsim.api.core.v01.network.Network;
+import org.matsim.contrib.ev.EvUnits;
 import org.matsim.contrib.ev.discharging.DrivingEnergyConsumptionEvent;
 import org.matsim.contrib.ev.discharging.DrivingEnergyConsumptionEventHandler;
+import org.matsim.core.controler.IterationCounter;
+import org.matsim.core.controler.OutputDirectoryHierarchy;
 import org.matsim.core.events.MobsimScopeEventHandler;
+import org.matsim.core.mobsim.framework.events.MobsimBeforeCleanupEvent;
+import org.matsim.core.mobsim.framework.listeners.MobsimBeforeCleanupListener;
+
+import com.google.inject.Inject;
 
 /**
  * @author Michal Maciejewski (michalm)
  */
-public class EnergyConsumptionCollector implements DrivingEnergyConsumptionEventHandler, MobsimScopeEventHandler {
+public class EnergyConsumptionCollector implements DrivingEnergyConsumptionEventHandler, MobsimScopeEventHandler, MobsimBeforeCleanupListener {
+
+	@Inject
+	private OutputDirectoryHierarchy controlerIO;
+	@Inject
+	private IterationCounter iterationCounter;
+	@Inject
+	private Network network;
 
 	private final Map<Id<Link>, Double> energyConsumptionPerLink = new HashMap<>();
 
@@ -41,7 +61,17 @@ public class EnergyConsumptionCollector implements DrivingEnergyConsumptionEvent
 		energyConsumptionPerLink.merge(event.getLinkId(), event.getEnergy(), Double::sum);
 	}
 
-	public Map<Id<Link>, Double> getEnergyConsumptionPerLink() {
-		return energyConsumptionPerLink;
+	@Override
+	public void notifyMobsimBeforeCleanup(MobsimBeforeCleanupEvent event) {
+		try (CSVPrinter csvPrinter2 = new CSVPrinter(Files.newBufferedWriter(
+				Paths.get(controlerIO.getIterationFilename(iterationCounter.getIterationNumber(), "evConsumptionPerLink.csv"))),
+				CSVFormat.DEFAULT.withDelimiter(';').withHeader("Link", "TotalConsumptionPerKm", "TotalConsumption"))) {
+			for (Map.Entry<Id<Link>, Double> e : energyConsumptionPerLink.entrySet()) {
+				csvPrinter2.printRecord(e.getKey(), (EvUnits.J_to_kWh(e.getValue())) / (network.getLinks().get(e.getKey()).getLength() / 1000.0),
+						EvUnits.J_to_kWh(e.getValue()));
+			}
+		} catch (IOException e) {
+			throw new RuntimeException(e);
+		}
 	}
 }
