--- contribs/ev/src/main/java/org/matsim/contrib/ev/stats/EvMobsimListener.java
@@ -1,88 +1,0 @@
-/* *********************************************************************** *
- * project: org.matsim.*
- *                                                                         *
- * *********************************************************************** *
- *                                                                         *
- * copyright       : (C) 2016 by the members listed in the COPYING,        *
- *                   LICENSE and WARRANTY file.                            *
- * email           : info at matsim dot org                                *
- *                                                                         *
- * *********************************************************************** *
- *                                                                         *
- *   This program is free software; you can redistribute it and/or modify  *
- *   it under the terms of the GNU General Public License as published by  *
- *   the Free Software Foundation; either version 2 of the License, or     *
- *   (at your option) any later version.                                   *
- *   See also COPYING, LICENSE and WARRANTY file                           *
- *                                                                         *
- * *********************************************************************** */
-
-package org.matsim.contrib.ev.stats;
-
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.util.Map;
-
-import org.apache.commons.csv.CSVFormat;
-import org.apache.commons.csv.CSVPrinter;
-import org.matsim.api.core.v01.Id;
-import org.matsim.api.core.v01.network.Link;
-import org.matsim.api.core.v01.network.Network;
-import org.matsim.contrib.ev.EvUnits;
-import org.matsim.core.controler.IterationCounter;
-import org.matsim.core.controler.OutputDirectoryHierarchy;
-import org.matsim.core.mobsim.framework.events.MobsimBeforeCleanupEvent;
-import org.matsim.core.mobsim.framework.listeners.MobsimBeforeCleanupListener;
-import org.matsim.core.utils.misc.Time;
-
-import com.google.inject.Inject;
-
-/*
- * created by jbischoff, 26.10.2018
- */
-public class EvMobsimListener implements MobsimBeforeCleanupListener {
-
-	@Inject
-	EnergyConsumptionCollector energyConsumptionCollector;
-	@Inject
-	ChargerPowerCollector chargerPowerCollector;
-	@Inject
-	OutputDirectoryHierarchy controlerIO;
-	@Inject
-	IterationCounter iterationCounter;
-	@Inject
-	Network network;
-
-	@Inject
-	EvMobsimListener() {
-	} // to make the constructor non-public.
-
-	@Override
-	public void notifyMobsimBeforeCleanup(MobsimBeforeCleanupEvent event) {
-		try (CSVPrinter csvPrinter = new CSVPrinter(
-				Files.newBufferedWriter(Paths.get(controlerIO.getIterationFilename(iterationCounter.getIterationNumber(), "chargingStats.csv"))),
-				CSVFormat.DEFAULT.withDelimiter(';')
-						.withHeader("ChargerId", "chargeStartTime", "chargeEndTime", "ChargingDuration", "xCoord", "yCoord",
-								"energyTransmitted_kWh"))) {
-			for (ChargerPowerCollector.ChargingLogEntry e : chargerPowerCollector.getLogList()) {
-				double energyKWh = Math.round(EvUnits.J_to_kWh(e.transmitted_Energy()) * 10.) / 10.;
-				csvPrinter.printRecord(e.charger().getId(), Time.writeTime(e.chargeStart()), Time.writeTime(e.chargeEnd()),
-						Time.writeTime(e.chargeEnd() - e.chargeStart()), e.charger().getCoord().getX(), e.charger().getCoord().getY(), energyKWh);
-			}
-		} catch (IOException e) {
-			throw new RuntimeException(e);
-		}
-
-		try (CSVPrinter csvPrinter2 = new CSVPrinter(Files.newBufferedWriter(
-				Paths.get(controlerIO.getIterationFilename(iterationCounter.getIterationNumber(), "evConsumptionPerLink.csv"))),
-				CSVFormat.DEFAULT.withDelimiter(';').withHeader("Link", "TotalConsumptionPerKm", "TotalConsumption"))) {
-			for (Map.Entry<Id<Link>, Double> e : energyConsumptionCollector.getEnergyConsumptionPerLink().entrySet()) {
-				csvPrinter2.printRecord(e.getKey(), (EvUnits.J_to_kWh(e.getValue())) / (network.getLinks().get(e.getKey()).getLength() / 1000.0),
-						EvUnits.J_to_kWh(e.getValue()));
-			}
-		} catch (IOException e) {
-			throw new RuntimeException(e);
-		}
-	}
-}
